完成主界面背景切换逻辑，解决第一次使用没有添加当前位置，完成设置界面，优化一些小问题

+++ app/src/main/java/com/fineweather/android/ui/MainAirQualityActivity.kt
@@ -58,7 +58,7 @@ class MainAirQualityActivity : AppCompatActivity() {
         }//设置空气质量和描述字体颜色
         Airquality_pm25.text=Airqualitydata.pm25.toInt().toString()
         Airquality_pm10.text=Airqualitydata.pm10.toInt().toString()
-        Airquality_SO2.text=Airqualitydata.so2.toString()
+        Airquality_SO2.text=Airqualitydata.so2.toInt().toString()
         Airquality_NO2.text=Airqualitydata.no2.toInt().toString()
         Airquality_O3.text=Airqualitydata.o3.toInt().toString()
         Airquality_CO.text=Airqualitydata.co.toString()
